fix(ci): python-full-coverage maintainability gate — split test files over 1500-line limit
The first-party line-limit guardrail (test_first_party_code_files_stay_under_line_limit)
failed because four test files exceeded MAX_FIRST_PARTY_FILE_LINES (1500):
test_status_part_001 (1580), test_health_part_001 (1517),
test_mcp_operator_surfaces_part_002 (1523), and test_orphan_resources (1818).

Rebalanced each over the existing _parts convention without changing any test
behavior: moved orphan/docker readiness cases into the sibling part_002 files,
added test_mcp_operator_surfaces_part_004, and split the standalone
test_orphan_resources.py into a _parts package. No assertions weakened.
