+++ tests/unit/api/test_health_parts/test_health_part_002.py
@@ -14,7 +14,7 @@ from __future__ import annotations
 
 import json
 from collections.abc import AsyncIterator
-from datetime import UTC, datetime
+from datetime import UTC, datetime, timedelta
 from pathlib import Path
 from typing import Any
 
@@ -329,3 +329,167 @@ async def test_readyz_terminal_workspace_with_live_container_reports_leak(
     assert example["workspace_id"] == workspace_id
     assert example["classification"] == "terminal"
     assert example["reason"] == "WORKSPACE_TERMINAL_LIVE_RUNTIME"
+
+
+@pytest.mark.unit
+async def test_readyz_docker_cli_missing_returns_503(
+    ready_app_and_client: tuple[Any, AsyncClient],
+) -> None:
+    """When the docker binary is absent, asyncio raises FileNotFoundError —
+    the readiness check must catch it and report an actionable reason."""
+    app, client = ready_app_and_client
+
+    class _DockerMissingRunner:
+        def __init__(self) -> None:
+            self.calls: list[list[str]] = []
+
+        async def run(
+            self,
+            args: list[str],
+            *,
+            input_bytes: bytes | None = None,
+            cwd: str | None = None,
+        ) -> CommandResult:
+            self.calls.append(list(args))
+            if args and args[0] == "docker":
+                raise FileNotFoundError(args[0])
+            return CommandResult(returncode=0, stdout="", stderr="")
+
+    app.state.command_runner = _DockerMissingRunner()
+
+    response = await client.get("/readyz")
+    assert response.status_code == 503
+    body = response.json()
+
+    cli_check = body["checks"]["docker_cli"]
+    assert cli_check["ok"] is False
+    assert cli_check["reason"] == "DOCKER_CLI_NOT_FOUND"
+
+    # Daemon, compose, and image checks all depend on docker — they should also fail
+    # (gracefully) rather than crash the request.
+    assert body["checks"]["docker_daemon"]["ok"] is False
+    assert body["checks"]["docker_compose"]["ok"] is False
+    assert body["checks"]["agent_runtime_image"]["ok"] is False
+
+
+@pytest.mark.unit
+async def test_readyz_docker_daemon_unreachable_returns_503(
+    ready_app_and_client: tuple[Any, AsyncClient],
+) -> None:
+    app, client = ready_app_and_client
+    runner = FakeCommandRunner()
+    runner.queue_result(stdout="Docker version 27.0.3, build abc\n")  # docker --version
+    runner.queue_result(  # docker info
+        returncode=1,
+        stderr="Cannot connect to the Docker daemon at unix:///var/run/docker.sock\n",
+    )
+    runner.queue_result(stdout="v2.29.2\n")
+    runner.queue_result(stdout="sha256:deadbeef\n")
+    app.state.command_runner = runner
+
+    response = await client.get("/readyz")
+    assert response.status_code == 503
+    daemon = response.json()["checks"]["docker_daemon"]
+    assert daemon["ok"] is False
+    assert daemon["reason"] == "DOCKER_DAEMON_UNREACHABLE"
+    assert "Cannot connect" in (daemon["detail"] or "")
+    orphan_check = response.json()["checks"]["orphan_resources"]
+    assert orphan_check["ok"] is True
+    assert orphan_check["status"] == "unavailable"
+    assert orphan_check["reason"] == "DOCKER_RESOURCE_SCAN_UNAVAILABLE"
+
+
+@pytest.mark.unit
+async def test_readyz_orphan_resources_present_returns_503(
+    ready_app_and_client: tuple[Any, AsyncClient],
+    engine: AsyncEngine,
+) -> None:
+    app, client = ready_app_and_client
+    workspace_id = await create_workspace(
+        engine,
+        status=WorkspaceStatus.completed,
+        updated_at=datetime.now(UTC) - timedelta(hours=200),
+    )
+    runner = FakeCommandRunner()
+    _queue_all_ok(runner)
+    runner.queue_result(
+        stdout=json.dumps(
+            {
+                "id": "abc",
+                "name": f"awf_{workspace_id}-agent-1",
+                "project": f"awf_{workspace_id}",
+                "service": "agent",
+                "state": "exited",
+                "status": "Exited (0)",
+            }
+        )
+        + "\n"
+    )
+    runner.queue_result(stdout="")
+    runner.queue_result(stdout="")
+    app.state.command_runner = runner
+
+    response = await client.get("/readyz")
+    body = response.json()
+
+    assert response.status_code == 503
+    assert body["status"] == "fail"
+    orphan_check = body["checks"]["orphan_resources"]
+    assert orphan_check["ok"] is False
+    assert orphan_check["reason"] == "ORPHAN_RESOURCES_PRESENT"
+    assert orphan_check["orphan_count"] == 1
+    assert orphan_check["cleanup_readiness"]["dry_run_only"] is True
+    assert orphan_check["examples"][0]["workspace_id"] == workspace_id
+
+
+@pytest.mark.unit
+async def test_readyz_orphan_resources_reflect_auto_cleanup_enabled(
+    ready_app_and_client: tuple[Any, AsyncClient],
+    engine: AsyncEngine,
+    tmp_path: Path,
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """With ``AWF_AUTO_CLEANUP_ORPHANS`` enabled the readiness summary must stop
+    advertising the dry-run-only posture so dashboards reflect that the worker reaps."""
+    app, client = ready_app_and_client
+    monkeypatch.setattr(
+        health_route,
+        "get_settings",
+        lambda: Settings(
+            _env_file=None,
+            host_home=str(tmp_path / "home"),
+            work_dir=str(tmp_path / "work"),
+            auto_cleanup_orphans=True,
+        ),
+    )
+    workspace_id = await create_workspace(
+        engine,
+        status=WorkspaceStatus.completed,
+        updated_at=datetime.now(UTC) - timedelta(hours=200),
+    )
+    runner = FakeCommandRunner()
+    _queue_all_ok(runner)
+    runner.queue_result(
+        stdout=json.dumps(
+            {
+                "id": "abc",
+                "name": f"awf_{workspace_id}-agent-1",
+                "project": f"awf_{workspace_id}",
+                "service": "agent",
+                "state": "exited",
+                "status": "Exited (0)",
+            }
+        )
+        + "\n"
+    )
+    runner.queue_result(stdout="")
+    runner.queue_result(stdout="")
+    app.state.command_runner = runner
+
+    response = await client.get("/readyz")
+    body = response.json()
+
+    orphan_check = body["checks"]["orphan_resources"]
+    assert orphan_check["reason"] == "ORPHAN_RESOURCES_PRESENT"
+    assert orphan_check["cleanup_readiness"]["dry_run_only"] is False
+    assert "Reaping is enabled" in orphan_check["cleanup_readiness"]["action"]
